% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BsProb1.R
\name{BsProb1}
\alias{BsProb1}
\title{Posterior Probabilities from Bayesian Screening
Experiments}
\usage{
BsProb1(X, y, p = 0.25, gamma = 2, max_int = 3, max_fac = ncol(X), top = 10)
}
\arguments{
\item{X}{Matrix. The design matrix.}

\item{y}{Vector. The response vector.}

\item{p}{Numeric. Prior probability assigned to active factors.}

\item{gamma}{Numeric. Variance inflation associated to active factors.}

\item{max_int}{Integer <= 3. Maximum order of interactions considered in the models.}

\item{max_fac}{Integer. Maximum number of factors included in the models.}

\item{top}{Integer. Number of models to keep with the highest posterior probability.}
}
\value{
A list with all the input and output parameters.

\item{X}{Matrix. The design matrix.}

\item{y}{Vector. The response vector.}

\item{n}{Integer. Number of runs.}

\item{col}{Integer. Number of columns in the design matrix.}

\item{max_int}{Integer <= 3. Maximum order of interactions considered in the models.}

\item{max_fac}{Integer. Maximum number of factors included in the models.}

\item{pi}{Numeric. Prior probability assigned to active factors.}

\item{gamma}{Numeric. Variance inflation associated with active factors.}

\item{top}{Integer. Number of models to keep with the highest posterior probability.}

\item{Prob_fac}{Data frame. Posterior probability for each factor.}

\item{Prob_mod}{Data frame. Posterior probability for each of the top models.}

\item{nfac_mod}{Vector. Number of active factors in each of the top models.}

\item{p_mod}{Vector. Posterior probability for each of the top models.}

\item{fac_mod}{Matrix. Active factors for each of the top models.}
}
\description{
Marginal factor posterior probabilities and model
posterior probabilities from designed screening
experiments are calculated according to Box and
Meyer's Bayesian procedure.
}
\examples{
#Example 1
library(BsMD2)
data("BM93e1")
X <- as.matrix(BM93e1[,2:6])
y <- BM93e1[,7]
drillAdvance.BsProb1 <- BsMD2::BsProb1(X, y, .25, 1.6, 3, 5)
plot(drillAdvance.BsProb1)
summary(drillAdvance.BsProb1)

#Example 2
data("BM93e2")
X <- as.matrix(BM93e2[,1:7])
y <- BM93e2[,8]
pp <- BsMD2::BsProb1(X, y, .25, 1.5, 3, 7)
plot(pp)
summary(pp)

#Example 3
data("BM93e3")
X16 <- as.matrix(BM93e3[1:16,2:9])
y16 <- BM93e3[1:16,10]
pp16 <- BsMD2::BsProb1(X16, y16, .25, 2, 3, 8)

X <- as.matrix(BM93e3[,1:9])
y <- BM93e3[,10]
pp <- BsProb1(X, y, .25, 2, 3, 4)


}
\references{
{
Box, G. E. P and R. D. Meyer (1986). "An Analysis for Unreplicated Fractional Factorials".
Technometrics. Vol. 28. No. 1. pp. 11{–}18.

Box, G. E. P and R. D. Meyer (1993). "Finding the Active Factors in Fractionated Screening Experiments".
Journal of Quality Technology. Vol. 25. No. 2. pp. 94{–}105.}
}
