% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lenth.R
\name{Lenth}
\alias{Lenth}
\title{Lenth's Plot of Effects}
\usage{
Lenth(fit)
}
\arguments{
\item{fit}{Object of class \link{lm}. Fitted model from lm.}
}
\value{
The function returns Lenth's plot. Dashed lines correspond to the margin of error (ME) and dotted lines to the simultaneous margin of error (SME).
}
\description{
Plot of the factor effects with significance levels based on robust estimation of contrast standard errors.
}
\examples{
#Example 6.2 Douglas C. Montgomery
#A <- rep(c(-1,1),8)
#B <- rep(c(-1,-1,1,1),4)
#C <- rep(c(rep(-1,4),rep(1,4)),2)
#D <- c(rep(-1,8), rep(1,8))
#rate <- c(45,71,48,65,68,60,80,65,43,100,45,104,75,86,70,96)
#fit <- lm(rate~A*B*C*D)
#Lenth(fit)
}
