% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BsProb1.R
\name{plot.BsProb1}
\alias{plot.BsProb1}
\title{Plot BsProb1 object}
\usage{
\method{plot}{BsProb1}(x, fac = TRUE, ...)
}
\arguments{
\item{x}{Object of class BsProb1.}

\item{fac}{Logical. If true, the plot of each factor's posterior probability is displayed.}

\item{...}{additional graphical parameters passed to plot.}
}
\value{
Plot of BsProb1 object
}
\description{
Plot BsProb1 object
}
