% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BsProb1.R
\name{print.BsProb1}
\alias{print.BsProb1}
\title{Print BsProb1 Object}
\usage{
\method{print}{BsProb1}(x, ...)
}
\arguments{
\item{x}{BsProb1 object.}

\item{...}{additional parameters}
}
\value{
Prints BsProb1 object
}
\description{
Print BsProb1 Object
}
