% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDopt.R
\name{print.MDopt}
\alias{print.MDopt}
\title{Print MDopt Object}
\usage{
\method{print}{MDopt}(x, ...)
}
\arguments{
\item{x}{MDopt object}

\item{...}{additional parameters}
}
\value{
Prints MDopt object
}
\description{
Print MDopt Object
}
