% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BsProb1.R
\name{summary.BsProb1}
\alias{summary.BsProb1}
\title{Summary for BsProb1 Object}
\usage{
\method{summary}{BsProb1}(object, ...)
}
\arguments{
\item{object}{BsProb1 Object}

\item{...}{additional parameters}
}
\value{
Summary for BsProb1 object.
}
\description{
Summary for BsProb1 Object
}
