% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDopt.R
\name{summary.MDopt}
\alias{summary.MDopt}
\title{Summary for MDopt Object}
\usage{
\method{summary}{MDopt}(object, ...)
}
\arguments{
\item{object}{MDopt Object}

\item{...}{additional parameters}
}
\value{
Summary for MDopt object.
}
\description{
Summary for MDopt Object
}
